refactor: moved all "service" type to "api"
there is no service and it is not related to spring or having dependency

Public API changed
