+++ cqp-storage-elasticsearch/src/main/java/com/quantori/cqp/storage/elasticsearch/ElasticsearchStorageConfiguration.java
@@ -8,13 +8,13 @@ import co.elastic.clients.transport.ElasticsearchTransport;
 import co.elastic.clients.transport.rest_client.RestClientTransport;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.quantori.cqp.api.StorageConfiguration;
-import com.quantori.cqp.api.service.MoleculesFingerprintCalculator;
-import com.quantori.cqp.api.service.MoleculesMatcher;
-import com.quantori.cqp.api.service.ReactionsFingerprintCalculator;
-import com.quantori.cqp.api.service.ReactionsMatcher;
-import com.quantori.cqp.api.service.StorageLibrary;
-import com.quantori.cqp.api.service.StorageMolecules;
-import com.quantori.cqp.api.service.StorageReactions;
+import com.quantori.cqp.api.MoleculesFingerprintCalculator;
+import com.quantori.cqp.api.MoleculesMatcher;
+import com.quantori.cqp.api.ReactionsFingerprintCalculator;
+import com.quantori.cqp.api.ReactionsMatcher;
+import com.quantori.cqp.api.StorageLibrary;
+import com.quantori.cqp.api.StorageMolecules;
+import com.quantori.cqp.api.StorageReactions;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpHost;
