+++ cqp-storage-elasticsearch/src/main/java/com/quantori/cqp/storage/elasticsearch/ElasticsearchStorageReactions.java
@@ -10,9 +10,9 @@ import co.elastic.clients.transport.ElasticsearchTransport;
 import com.quantori.cqp.api.model.Flattened;
 import com.quantori.cqp.api.model.Library;
 import com.quantori.cqp.api.model.upload.ReactionUploadDocument;
-import com.quantori.cqp.api.service.ReactionsFingerprintCalculator;
-import com.quantori.cqp.api.service.ReactionsMatcher;
-import com.quantori.cqp.api.service.StorageReactions;
+import com.quantori.cqp.api.ReactionsFingerprintCalculator;
+import com.quantori.cqp.api.ReactionsMatcher;
+import com.quantori.cqp.api.StorageReactions;
 import com.quantori.cqp.api.util.FingerPrintUtilities;
 import com.quantori.cqp.core.model.ExactParams;
 import com.quantori.cqp.core.model.ItemWriter;
